server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
paypal-checkout-serversdk
pydantic
